requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
PyYAML>=6.0
joblib>=1.3.0
matplotlib>=3.7.0
python-telegram-bot>=21.0
pytest>=7.0.0
pyarrow>=10.0.0
yfinance>=0.2.18
seaborn>=0.13.0
alpha_vantage>=2.3.1
python-binance>=1.0.19
ccxt>=4.0.0
aiohttp>=3.8.0
websocket-client>=1.0.0
requests>=2.25.0

# Optimization dependencies
optuna>=3.4.0
optuna-dashboard>=0.13.0
scikit-optimize>=0.9.0
plotly>=5.15.0
kaleido>=0.2.1
torch>=2.0.0
torchvision>=0.15.0

# Optimization
pymoo>=0.6.0
psutil>=5.9.0